# Improved runtime performance