--- a/src/getBinDirectory.ts
+++ b/src/getBinDirectory.ts
@@ -1,10 +1,8 @@
 import { exec } from "child_process";
-import path from "path";
 
 export function getBinDirectory() {
   return new Promise<string>((resolve, reject) => {
-    const programToFind = "exec-if-exists";
-    const command = `npm ls --parseable ${programToFind}`;
+    const command = `npm bin`;
     exec(
       command,
       {
@@ -12,26 +10,16 @@ export function getBinDirectory() {
       },
       (error, stdout, stderr) => {
         const code = error?.code;
-        if (code == null || code === 0) {
-          if (stdout.length === 0) {
-            reject(
-              `Error: ${programToFind} needs to be installed in the local directory to run`
-            );
-            return;
-          }
-
-          const index = stdout.lastIndexOf(programToFind);
-          const lastLineStart = stdout.trim().lastIndexOf("\n") + 1;
-          const binPath = path.join(
-            stdout.substring(lastLineStart, index),
-            ".bin"
-          );
-          resolve(binPath);
-          return;
-        } else {
-          reject(`Error: npm ls returned code ${code}`);
-          return;
+        if (code != null && code !== 0) {
+          return reject(`Error: npm bin returned code ${code}`);
         }
+
+        if (stdout.length === 0) {
+          return reject(`Error: npm bin did not return a value`);
+        }
+
+        const binPath = stdout.trim();
+        return resolve(binPath);
       }
     );
   });