--- a/src/programInBin.ts
+++ b/src/programInBin.ts
@@ -4,22 +4,20 @@ export function programInBin(binDir: string, programName: string) {
   return new Promise<boolean>((resolve, reject) => {
     fs.readdir(binDir, (err, files) => {
       if (err != null) {
-        reject(`Error ${err.message ?? err}`);
-        return;
+        return reject(`Error ${err.message ?? err}`);
       }
 
       if (files == null || files.length === 0) {
-        reject("Error: No node programs found. Did you run npm install?");
-        return;
+        return reject(
+          "Error: No node programs found. Did you run npm install?"
+        );
       }
 
       if (files.indexOf(programName) !== -1) {
-        resolve(true);
-        return;
+        return resolve(true);
       }
 
-      resolve(false);
-      return;
+      return resolve(false);
     });
   });
 }