Old perf:
getBinDirectory: 1.243s
programInBin: 1.353ms
runNpx: 4.564ms

New perf:
getBinDirectory: 780.195ms
programInBin: 1.185ms
runNpx: 5.212ms